refactor: extract shared test helpers to TestHelpers.swift
DRY
