+++ ios/SookeCommunityTests/ViewModels/BusinessDetailViewModelTests.swift
@@ -12,13 +12,6 @@ import Testing
 @Suite("Business Detail ViewModel Tests")
 @MainActor
 struct BusinessDetailViewModelTests {
-    func makeTestClient() -> APIClient {
-        let config = URLSessionConfiguration.ephemeral
-        config.protocolClasses = [MockURLProtocol.self]
-        let session = URLSession(configuration: config)
-        return APIClient(baseURL: "http://localhost:8080", session: session)
-    }
-
     func makeBusinessDetailsJSON(
         id: Int64 = 1,
         name: String = "Test Cafe",
@@ -28,13 +21,7 @@ struct BusinessDetailViewModelTests {
         {"id": \(id), "name": "\(name)", "slug": "\(slug)", "description": "A test cafe", "category_name": "Food", "category_slug": "food", "address": "123 Main St", "latitude": 48.37, "longitude": -123.72, "phone": null, "email": null, "website": null, "hours": [{"day_of_week": 1, "open_time": "09:00:00", "close_time": "17:00:00", "is_closed": false}], "menus": []}
         """.data(using: .utf8)!
     }
-
-    func makeErrorJSON(code: String = "not_found", message: String = "Business not found") -> Data {
-        """
-        {"error": {"code": "\(code)", "message": "\(message)"}}
-        """.data(using: .utf8)!
-    }
-
+    
     @Test func fetchesBusinessDetails() async throws {
         MockURLProtocol.reset()
         MockURLProtocol.mockResponseData = makeBusinessDetailsJSON()
