+++ ios/SookeCommunityTests/ViewModels/BusinessListViewModelTests.swift
@@ -12,52 +12,6 @@ import Testing
 @Suite("Business List ViewModel Tests")
 @MainActor
 struct BusinessListViewModelTests {
-    func makeTestClient() -> APIClient {
-        let config = URLSessionConfiguration.ephemeral
-        config.protocolClasses = [MockURLProtocol.self]
-        let session = URLSession(configuration: config)
-        return APIClient(baseURL: "http://localhost:8080", session: session)
-    }
-    
-    func makeCategoryJSON(
-        categories: [(id: Int64, name: String, slug: String)] = [
-            (1, "Food", "food"),
-            (2, "Retail", "retail")
-        ]
-    ) -> Data {
-        let items = categories.map { c in
-            "{\"id\": \(c.id), \"name\": \"\(c.name)\", \"slug\": \"\(c.slug)\"}"
-        }.joined(separator: ",")
-                                                                                                                              
-        return """
-          {"items": [\(items)]}
-          """.data(using: .utf8)!
-    }
-    
-    func makeErrorJSON(code: String = "server_error", message: String = "Internal Server Error") -> Data {
-          """
-          {"error": {"code": "\(code)", "message": "\(message)"}}
-          """.data(using: .utf8)!
-    }
-    
-    func makePaginatedBusinessJSON(
-        businesses: [(id: Int64, name: String, slug: String, categoryName: String, categorySlug: String)] = [
-            (1, "Test Cafe", "test-cafe", "Food", "food")
-        ],
-        page: Int = 1,
-        perPage: Int = 20,
-        totalItems: Int? = nil,
-        totalPages: Int = 1
-    ) -> Data {
-        let items = businesses.map { b in
-            "{\"id\": \(b.id), \"name\": \"\(b.name)\", \"slug\": \"\(b.slug)\", \"description\": null, \"category_name\": \"\(b.categoryName)\", \"category_slug\": \"\(b.categorySlug)\", \"address\": \"123 Main St\", \"latitude\": 48.37, \"longitude\": -123.72, \"phone\": null, \"email\": null, \"website\": null}"
-        }.joined(separator: ",")
-
-        return """
-        {"items": [\(items)], "pagination": {"page": \(page), "per_page": \(perPage), "total_items": \(totalItems ?? businesses.count), "total_pages": \(totalPages)}}
-        """.data(using: .utf8)!
-    }
-    
     @Test func fetchesBusinesses() async throws {
         MockURLProtocol.reset()
         MockURLProtocol.mockResponseData = makePaginatedBusinessJSON()
@@ -65,7 +19,7 @@ struct BusinessListViewModelTests {
         vm.apiClient = makeTestClient()
         await vm.fetchBusinesses()
 
-        #expect(vm.items.count == 1)
+        #expect(vm.items.count == 3)
         #expect(vm.items.first?.name == "Test Cafe")
         #expect(vm.isLoading == false)
     }
@@ -110,7 +64,7 @@ struct BusinessListViewModelTests {
         vm.apiClient = makeTestClient()
         await vm.fetchCategories()
 
-        #expect(vm.categories.count == 2)
+        #expect(vm.categories.count == 3)
         #expect(vm.categories[0].name == "Food")
         #expect(vm.categories[1].name == "Retail")
         #expect(vm.isLoading == false)
@@ -135,9 +89,8 @@ struct BusinessListViewModelTests {
         await vm.fetchBusinesses()
         
         let url = MockURLProtocol.lastRequest?.url?.absoluteString
-        print("Requested URL: \(url ?? "nil")")
         
-        #expect(vm.items.count == 1)
+        #expect(vm.items.count == 3)
         #expect(vm.items.first?.name == "Test Cafe")
         #expect(vm.isLoading == false)
         #expect(url?.contains("search=Cafe") == true)
@@ -154,7 +107,7 @@ struct BusinessListViewModelTests {
 
         let url = MockURLProtocol.lastRequest?.url?.absoluteString
 
-        #expect(vm.items.count == 1)
+        #expect(vm.items.count == 3)
         #expect(vm.items.first?.name == "Test Cafe")
         #expect(vm.isLoading == false)
         #expect(url?.contains("category=food") == true)
@@ -173,7 +126,7 @@ struct BusinessListViewModelTests {
         let url = MockURLProtocol.lastRequest?.url?.absoluteString
         
         #expect(vm.selectedCategory == testCategory)
-        #expect(vm.items.count == 1)
+        #expect(vm.items.count == 3)
         #expect(vm.items.first?.name == "Test Cafe")
         #expect(vm.isLoading == false)
         #expect(url?.contains("category=food") == true)
